// DF/src/erd/vgph_system_setup/dto/Updatevgph_system_setup.dto.ts
import { Prisma } from '@prisma/client';
import { IsEnum } from 'class-validator';
import { ApiProperty } from '@nestjs/swagger';


export class  Updatevgph_system_setupDto {
        @ApiProperty()
        setup_code?: string;
        @ApiProperty()
        product_code?: string;
        @ApiProperty()
        interface_product?: string;
        @ApiProperty()
        category?: string;
        @ApiProperty()
        sub_category?: string;
        @ApiProperty()
        purpose?: string;
        @ApiProperty()
        setup_value?: Prisma.InputJsonValue;
        @ApiProperty()
        tenant_id?: string;


























        @ApiProperty({
            type: `string`,
            format: `date-time`,
        })
        trs_created_date?: Date;
        @ApiProperty()
        trs_created_by?: string;
        @ApiProperty({
            type: `string`,
            format: `date-time`,
        })
        trs_modified_date?: Date;
        @ApiProperty()
        trs_modified_by?: string;
        @ApiProperty()
        trs_status?: string;
        @ApiProperty()
        trs_next_status?: string;
        @ApiProperty()
        trs_process_id?: string;
        @ApiProperty()
        trs_access_profile?: string;
        @ApiProperty()
        trs_org_grp_code?: string;
        @ApiProperty()
        trs_org_code?: string;
        @ApiProperty()
        trs_role_grp_code?: string;
        @ApiProperty()
        trs_role_code?: string;
        @ApiProperty()
        trs_ps_grp_code?: string;
        @ApiProperty()
        trs_ps_code?: string;
        @ApiProperty()
        trs_sub_org_code?: string;
        @ApiProperty()
        trs_sub_org_grp_code?: string;

}